Drop odataQueryHandler
